Bug 1881324
git-svn-id: https://clipsrules.svn.sourceforge.net/svnroot/clipsrules@12 b3187f35-aa44-0410-a219-a43ecd7132bf

+++ core/facthsh.c
@@ -149,7 +149,7 @@ globle intBool RemoveHashedFact(
   void *theEnv,
   struct fact *theFact)
   {
-   int hashValue;
+   unsigned long hashValue;
    struct factHashEntry *hptr, *prev;
 
    hashValue = HashFact(theFact);
